requirements.txt
requests==2.31.0
feedparser==6.0.11
pandas==2.1.4
numpy==1.24.3
chromadb==0.4.14
posthog<4
python-dotenv==1.0.0
python-dateutil==2.9.0.post0
apscheduler==3.10.4
pydantic==2.5.0
httpx>=0.27,<0.28
beautifulsoup4==4.12.2
PyYAML==6.0.2
Jinja2==3.1.6
lxml==4.9.3
openai==1.3.0
yfinance>=0.2.36
langchain==0.1.0
sentence-transformers==2.2.2
torch==2.1.1
transformers==4.35.2
huggingface-hub==0.19.4
scikit-learn==1.3.2
# Análise Econométrica Profissional
statsmodels==0.14.0
scipy==1.11.4
arch==6.2.0
prophet==1.1.5
# Banco de Dados
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
# Visualização
plotly==5.18.0
matplotlib==3.8.2
